Fix minor visual bug with ReorderableListViews.
This project is now running Flutter 2.0, which changes how
ReorderableListViews work. Previously, ListViewCards were given an
icon to show that they could be reordered on their right side.
However, in Flutter 2.0, ReorderableListViews do that automatically
to all of its children. So, this feature was removed, since its no
longer necessary.