Store info about logged in user
When the user logs in or registers and the HomeScreen loads, it will
automatically save info about the logged in user in a static instance
of StrydeUserStorage. This information about the user can be accessed
from any screen, making it a helpful utility for getting info about the
user that's currently logged in without querying the server ever time
info is needed.