Create & implement a SearchableListView
Create the AllExerciseListScreen, which enables users to view all
possible exercises. When an exercise is clicked, it will add it to the
current workout. However, this is currently just a facade. We still
need to get all of a user's exercises from the database to be
displayed, add the functionality of adding it to the workout when it's
clicked, and finally save the workout after all of that has been done.