Fix bug that chose the wrong exercise sometimes
Fix bug on AllExerciseListScreen that caused the wrong exercise to be
selected sometimes. When the user searched for an exercise, and clicked
on a searched exercise, the list would be smaller. However, it returned
the index of the exercise at the original position, not the index of
the searched exercise. This caused the wrong exercise to be selected
sometimes when users searched for an exercise to add.